GameLogic

Description

This module contains the Brain of the Game. It allows input to be checked using the MapApi Lib and it allows the scoring of answers.

Documentation

getRandomChar

Arguments

:: IO Char
Random uppercase char

Generates a random uppercase Char

checkRiver

Arguments

:: [String]
List of all answers given

-> IO [(Int, String)]
Score and likelihood of every answer

Checks and scores a list of answers for the river type

checkCountry

Arguments

:: [String]
List of all answers given

-> IO [(Int, String)]
Score and likelihood of every Answer

Checks and scores a list of answers for the country type

checkCity

Arguments

:: [String]
List of all answers given

-> IO [(Int, String)]
Score and likelihood of every Answer

Checks and scores a list of answers for the city type

evaluateAnswer

Arguments

:: String
The entity type of the answer CountryRegion | River | PopulatedPlace

-> [String]
List of all Answers to determine score multiplier

-> String
The given Answer which will be scored

-> IO (Int, String)
Score and likelihood

Evaluates how likely a given answer is to be correct and how many points it should be worth Uses the Rest API to get all Locations matching the Answer.

scoreAnswer

Arguments

:: String
Entity Type which the location should have

-> [Location]
List of locations matching the query

-> String
Query / Answer to be scored

-> [String]
All Answers used for score multiplier

-> (Int, String)
Score and likelihood

Score a answer based on given locations

bestMatch

Arguments

:: String
Entity type which the result MUST match

-> [Location]
List of all Locations

-> Maybe Location
Best Match for given type, nothing if no result for given type was found

Filter a sorted List of locations for the best match in a given entity Type

isType

Arguments

:: String
Name of the type expected

-> Location
Location to be checked

-> Bool
true if entityType is as expected

determine if Location type matches

calcBonusFactor

Arguments

:: String
Answer for which the multiplier is calculated

-> [String]
List of all Answers (including the one being evaluated)

-> Int
Score multiplier (1 | 2 | 4)

Calculates a score multiplier for a given Answer 4: if only one non empty answer was given. (No one else answered in that category) 2: The answer is unique. (There where other answers, but not equal answers) 1: The answer was given multiple times. Note: Empty answers will get score multipliers too, but since they will receive 0 points the multiplier does not matter

currentScores :: [Int]