GuiMain

Description

This module builds up the components and design of the GUI. The GUI is build with the threepenny-gui library (http:/hackage.haskell.orgpackage/threepenny-gui) which uses HTML like statements in haskell to build up a GUI in a browser window which is hostet on localhost.

Documentation

guiMain :: IO ()

setup :: Window -> UI ()

Builds up the user window

mkDisplay :: UI Element

This defines the content inside the user window