MapApi

Description

This module handles accessing the Bing Maps api to return Locations for a given Query.

Documentation

data Location

Location describing a rest request result

Constructors

Location

Fields

name :: String

Name of the location found

entityType :: String

Entitiy Type of the location (e.g. River)

confidence :: String

how confident the Api was that this is the correct result (low, medium or high)

getLocation

Arguments

:: String
Query to be searched

-> IO [Location]
List of all locations returned from the api. Can be empty

Get a list of all matches for the given query string